// lib/appwrite/types.ts
import type { Models } from 'node-appwrite';

// This file is auto-generated by the Appwrite CLI. 
// You can regenerate it by running `appwrite types -l ts types.ts`.

export enum VoteType {
    UPVOTE = "upvote",
    DOWNVOTE = "downvote"
}

export enum InteractionActionType {
    QUESTION = "question",
    ANSWER = "answer"
}

export type AppUser = Models.Row & {
    name: string;
    username: string;
    email: string;
    bio: string | null;
    image: string | null;
    location: string | null;
    portfolio: string | null;
    reputation: number;
    answers: Answer[];
    questions: Question[];
    votes: Vote[];
    collection: Collection[];
    interaction: Interaction[];
    provider: string | null;
}

export type Question = Models.Row & {
    author: AppUser;
    title: string;
    content: string;
    views: number;
    upvotes: number;
    downvotes: number;
    answersCount: number;
    answers: Answer[];
    votes: Vote[];
    collection: Collection[];
    tags: QuestionsTags[];
}

export type Tag = Models.Row & {
    title: string;
    questionsCount: number;
    questions: QuestionsTags[];
}

export type QuestionsTags = Models.Row & {
    question: Question;
    tag: Tag;
}

export type Answer = Models.Row & {
    author: AppUser;
    question: Question;
    content: string;
    upvotes: number;
    downvotes: number;
    votes: Vote[];
}

export type Vote = Models.Row & {
    author: AppUser;
    question: Question;
    answer: Answer;
    type: VoteType;
}

export type Collection = Models.Row & {
    author: AppUser;
    question: Question;
}

export type Interaction = Models.Row & {
    user: AppUser;
    action: string;
    actionId: string;
    actionType: InteractionActionType;
}